decoder/requirements.txt
lz4>=4.0.0
Pillow>=9.0.0
texture2ddecoder>=1.0.0
